requirements_test.txt
coverage
mock
flake8
tox
codecov
pre-commit
freezegun
djangorestframework>3.14
